requirements.txt
# Core Deep Learning (Latest Versions)
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Modern Computer Vision
ultralytics>=8.0.0                 # YOLOv8/v11 
mmcv>=2.0.0                        # Modern CV framework
mmpose>=1.2.0                      # Advanced pose estimation
segment-anything-py>=1.0.0         # Meta's SAM for segmentation

# Pose Estimation Upgrade
mediapipe>=0.10.0                  # Latest MediaPipe
opencv-python>=4.8.0
numpy>=1.24.0

# Modern ML Tools
transformers>=4.35.0               # For Vision Transformers
timm>=0.9.0                        # Latest vision models
lightning>=2.1.0                   # Modern PyTorch Lightning

# Tracking and Temporal Analysis
filterpy>=1.4.5                    # Kalman filtering
scipy>=1.11.0

# Data and Visualization
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
pillow>=10.0.0

# Modern UI (Optional - can add later)
gradio>=4.0.0                      # Better than Flask for demos
streamlit>=1.28.0                  # Dashboard alternative

# Development Tools
wandb>=0.16.0                      # Experiment tracking
tqdm>=4.66.0
pyyaml>=6.0.1
hydra-core>=1.3.0                  # Configuration management

# Google Colab Integration
google-colab
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client

# Performance Optimization
onnx>=1.15.0                       # Model optimization
tensorrt>=8.6.0                    # NVIDIA optimization (if available)